More bug fixes and updates
